app/requirements_dev.txt
black==19.10b0
flake8===3.7.9
pytest-cov==2.8.1
pytest-django==3.9.0
pytest==5.4.1
isort==4.3.21

-r requirements.txtdc exec